--- packages/api/src/app.ts
@@ -1,68 +1,0 @@
-// EXPRESS
-import express, { Express } from 'express';
-import cors from 'cors';
-import bodyParser from 'body-parser';
-// PASSPORT
-import passport from 'passport';
-import './middlewares/passport';
-// SWAGGER
-import swaggerJsdoc from 'swagger-jsdoc';
-import swaggerUi from 'swagger-ui-express';
-// GRAPHQL
-import { ApolloServer } from 'apollo-server-express';
-import { buildContext } from 'graphql-passport';
-import { typeDefs } from './graphQl/typeDefs';
-import { resolvers } from './graphQl/resolvers';
-// INTERNALS
-import Routes from './routes';
-import User from './entity/User';
-
-const app: Express = express();
-
-app.use(passport.initialize());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
-app.use(cors());
-
-app.use('/api', Routes);
-
-app.get('/', (req, res) => res.status(200).end('Type /api to use it'));
-
-// Swagger set up
-const options = {
-	swaggerDefinition: {
-		openapi: '3.0.0',
-		info: {
-			title: 'API eBet - Documentation',
-			version: '1.0.0',
-		},
-		servers: [
-			{
-				url: `http://localhost:${process.env.PORT || 3000}/api/`,
-			},
-		],
-	},
-	apis: [
-		'./src/entity/User.ts',
-		'./src/controllers/AuthController.ts',
-		'./src/controllers/UserController.ts',
-	],
-};
-const specs = swaggerJsdoc(options);
-app.use('/doc', swaggerUi.serve);
-app.get(
-	'/doc',
-	swaggerUi.setup(specs, {
-		explorer: true,
-	}),
-);
-
-// GraphQL startup
-const server = new ApolloServer({
-	typeDefs,
-	resolvers,
-	context: ({ req, res }) => buildContext({ req, res, User }),
-});
-server.applyMiddleware({ app });
-
-export default app;
