+++ packages/api/src/routes/auth.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import AuthController from '../controllers/AuthController';
+import AuthController from '../controllers/rest/AuthController';
 
 const router: Router = Router();
