+++ packages/api/src/routes/index.ts
@@ -1,12 +1,14 @@
+// EXPRESS
 import { Router } from 'express';
+// PASSPORT
+import passport from 'passport';
+// INTERNALS
 import auth from './auth';
 import secured from './secured/';
-import passport from 'passport';
 
 const routes: Router = Router();
 
 routes.use('/auth', auth);
-
 routes.use('/', passport.authenticate('jwt', { session: false }), secured);
 
 export default routes;
