+++ packages/api/tests/suites/authRoutesSuite.ts
@@ -1,7 +1,7 @@
 import supertest from 'supertest';
 import request from 'superagent';
 // INTERNALS
-import User from '../../src/entity/User';
+import User from '../../src/database/models/User';
 import AuthenticateService from '../../src/services/AuthServices';
 
 const authRealTest = (server: supertest.SuperTest<supertest.Test>) =>
