+++ packages/api/tsconfig.json
@@ -12,7 +12,7 @@
 		"allowSyntheticDefaultImports": true,
 		"emitDecoratorMetadata": true,
 		"experimentalDecorators": true,
-		"outDir": ".dist"
+		"outDir": ".dist",
 	},
 	"files": ["types.d.ts"],
 	"include": ["src", "types.d.ts"],
